ok, kind of working
