--- src/main/java/com/pbuczek/pf/security/UserDetails.java
@@ -1,58 +1,0 @@
-package com.pbuczek.pf.security;
-
-import com.pbuczek.pf.user.User;
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-
-import java.util.Collection;
-import java.util.List;
-
-@Data
-public class UserDetails implements org.springframework.security.core.userdetails.UserDetails {
-
-    private Integer userName;
-    private String password;
-    private List<GrantedAuthority> authorities;
-
-    public UserDetails(User user) {
-        userName = user.getId();
-        password = user.getPassword();
-        authorities = List.of(new SimpleGrantedAuthority(user.getType().toString()));
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return String.valueOf(userName);
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
