--- src/main/java/com/pbuczek/pf/security/UserDetailsService.java
@@ -1,20 +1,0 @@
-package com.pbuczek.pf.security;
-
-import com.pbuczek.pf.user.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {
-
-    @Autowired
-    private UserRepository userRepo;
-
-    @Override
-    public org.springframework.security.core.userdetails.UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepo.findByUsername(username)
-                .map(UserDetails::new)
-                .orElseThrow(() -> new UsernameNotFoundException("No user found"));
-    }
-}
